+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Overbound
 ## Version 1.1
 - Fixed localization problem for Overwrite popup
+- Added confirmation to delete keybind
 - Added option to load bind once on startup
 
 ## Version 1.0